chore: improve code organization
